+++ beacon-chain/sync/validate_attester_slashing_test.go
@@ -3,7 +3,7 @@ package sync
 import (
 	"bytes"
 	"context"
-	"math/rand"
+	"crypto/rand"
 	"reflect"
 	"testing"
 	"time"